feat: make HorizontalScroller reusable
