+++ src/components/common/molecules/HorizontalScroller/index.tsx
@@ -1,69 +1,21 @@
-import { useRef } from 'react';
+import { ReactNode, useRef } from 'react';
 
 import {
   ItemWrapper,
-  StyledLongButton,
   Wrapper,
   StyledRightArrow,
   StyledLeftArrow,
 } from './styled';
 
-const userCultList = [
-  {
-    name: 'The High Collective of Shoe1',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe2',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe3',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe4',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe5',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe6',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe7',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe7',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe7',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe7',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe7',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe7',
-    link: 'https://google.com',
-  },
-  {
-    name: 'The High Collective of Shoe7',
-    link: 'https://google.com',
-  },
-];
+type Props<T> = {
+  data: T[];
+  renderItem: (item: T) => ReactNode;
+};
 
-const HorizontalScroller = () => {
+const HorizontalScroller = <T extends object>({
+  data,
+  renderItem,
+}: Props<T>) => {
   const scrollerRef = useRef<HTMLDivElement>(null);
 
   const scrollRight = () => {
@@ -82,17 +34,7 @@ const HorizontalScroller = () => {
     <Wrapper>
       <StyledLeftArrow onClick={scrollLeft} />
       <ItemWrapper ref={scrollerRef}>
-        {userCultList.map((item) => {
-          const { name, link } = item;
-          return (
-            <StyledLongButton
-              key={name}
-              onClick={() => window.open(link, '_self')}
-            >
-              {name}
-            </StyledLongButton>
-          );
-        })}
+        {data.map((item) => renderItem(item))}
       </ItemWrapper>
       <StyledRightArrow onClick={scrollRight} />
     </Wrapper>
