+++ src/components/common/molecules/HorizontalScroller/styled.tsx
@@ -1,11 +1,8 @@
 import { BiRightArrow, BiLeftArrow } from 'react-icons/bi';
 import styled from 'styled-components';
 
-import { LongButton } from 'src/components/common/atoms';
-
 const Wrapper = styled.div`
   position: relative;
-  // border: 1px solid ${({ theme }) => theme.colors.grey.deepmist};
   padding: 1rem;
   width: 37.6%;
   margin: auto;
@@ -24,10 +21,6 @@ const ItemWrapper = styled.div`
     -webkit-appearance: none;
     height: 0;
   }
-`;
-
-const StyledLongButton = styled(LongButton)`
-  margin-right: 10px;
 `;
 
 const StyledRightArrow = styled(BiRightArrow)`
@@ -59,10 +52,4 @@ const StyledLeftArrow = styled(BiLeftArrow)`
   }
 `;
 
-export {
-  ItemWrapper,
-  StyledLongButton,
-  Wrapper,
-  StyledRightArrow,
-  StyledLeftArrow,
-};
+export { ItemWrapper, Wrapper, StyledRightArrow, StyledLeftArrow };
